requirements.txt
SpeechRecognition==3.10.0
wikipedia==1.4.0
requests==2.31.0
gTTS==2.3.2
pygame==2.5.2
psutil==5.9.5
schedule==1.2.0
newsapi-python==0.2.6
python-dotenv==1.0.0
aiohttp==3.9.1
SQLAlchemy==2.0.23
aiosqlite==0.19.0
pydantic==2.5.2
tenacity==8.2.3
structlog==23.2.0 